Check if path is not empty before deleting signature

+++ src/AppBundle/Service/FileUploader.php
@@ -74,6 +74,10 @@ class FileUploader
 
     public function deleteSignature(string $path)
     {
+        if (empty($path)) {
+            return;
+        }
+
         $fileName = $this->getFileNameFromPath($path);
 
         $this->deleteFile("$this->signatureFolder/$fileName");
